+++ splunk_otel/symbols.py
@@ -14,14 +14,11 @@
 
 _DEFAULT_SERVICE_NAME = "unnamed-python-service"
 _DEFAULT_EXPORTERS = "otlp"
-_DEFAULT_JAEGER_ENDPOINT = "http://localhost:9080/v1/trace"
 _DEFAULT_MAX_ATTR_LENGTH = 12000
 _DEFAULT_SPAN_LINK_COUNT_LIMIT = 1000
 
 _EXPORTER_OTLP = "otlp"
 _EXPORTER_OTLP_GRPC = "otlp_proto_grpc"
-_EXPORTER_JAEGER_THRIFT = "jaeger_thrift"
-_EXPORTER_JAEGER_SPLUNK = "jaeger-thrift-splunk"
 _DEFAULT_OTEL_SERVICE_NAME = "unknown_service"
 
 _SPLUNK_DISTRO_VERSION_ATTR = "splunk.distro.version"
@@ -34,8 +31,6 @@ E.g. `OTEL_SERVICE_NAME="<YOUR_SERVICE_NAME_HERE>"`"""
 _KNOWN_EXPORTER_PACKAGES = {
     _EXPORTER_OTLP: "opentelemetry-exporter-otlp-proto-grpc",
     _EXPORTER_OTLP_GRPC: "opentelemetry-exporter-otlp-proto-grpc",
-    _EXPORTER_JAEGER_THRIFT: "opentelemetry-exporter-jaeger-thrift",
-    _EXPORTER_JAEGER_SPLUNK: "opentelemetry-exporter-jaeger-thrift",
 }
 
 _LIMIT_UNSET_VALUE = ""